+++ src/main/java/appland/editor/AppMapIconProvider.java
@@ -2,6 +2,7 @@ package appland.editor;
 
 import appland.Icons;
 import appland.files.AppMapFiles;
+import appland.milestones.UserMilestonesEditorProvider;
 import com.intellij.ide.FileIconProvider;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.vfs.VirtualFile;
@@ -15,6 +16,9 @@ public class AppMapIconProvider implements FileIconProvider {
     @Nullable
     public Icon getIcon(@NotNull VirtualFile file, int flags, @Nullable Project project) {
         if (AppMapFiles.isAppMap(file)) {
+            return Icons.APPMAP_FILE;
+        }
+        if (UserMilestonesEditorProvider.isQuickstartFile(file)) {
             return Icons.APPMAP_FILE;
         }
         return null;