+++ src/main/java/appland/lang/agents/InstallResult.java
@@ -1,0 +1,5 @@
+package appland.lang.agents;
+
+public enum InstallResult {
+    None, Upgraded, Installed;
+}