+++ src/main/java/appland/toolwindow/AppMapWindowPanel.java
@@ -4,6 +4,8 @@ import appland.AppMapBundle;
 import appland.actions.StartAppMapRecordingAction;
 import appland.actions.StopAppMapRecordingAction;
 import appland.files.AppMapFiles;
+import appland.milestones.UserMilestonesEditorProvider;
+import appland.toolwindow.milestones.UserMilestonesPanel;
 import com.intellij.ide.util.treeView.NodeDescriptor;
 import com.intellij.openapi.Disposable;
 import com.intellij.openapi.actionSystem.*;
@@ -21,6 +23,7 @@ import com.intellij.psi.PsiManager;
 import com.intellij.ui.DocumentAdapter;
 import com.intellij.ui.ScrollPaneFactory;
 import com.intellij.ui.SearchTextField;
+import com.intellij.ui.SimpleTextAttributes;
 import com.intellij.ui.content.Content;
 import com.intellij.ui.tree.AsyncTreeModel;
 import com.intellij.ui.tree.StructureTreeModel;
@@ -72,6 +75,7 @@ public class AppMapWindowPanel extends SimpleToolWindowPanel implements DataProv
 
         setToolbar(createToolBar(appMapModel));
         setContent(ScrollPaneFactory.createScrollPane(tree));
+        add(createUserMilestonesPanel(), BorderLayout.SOUTH);
 
         // refresh when dumb mode changes
         var busConnection = project.getMessageBus().connect(this);
@@ -140,6 +144,10 @@ public class AppMapWindowPanel extends SimpleToolWindowPanel implements DataProv
     private SimpleTree createTree(@NotNull Disposable disposable, StructureTreeModel<AppMapTreeModel> treeModel) {
         var tree = new SimpleTree(new AsyncTreeModel(treeModel, true, disposable));
         tree.getEmptyText().setText(AppMapBundle.get("toolwindow.appmap.emptyText"));
+        tree.getEmptyText().appendSecondaryText(
+                AppMapBundle.get("toolwindow.appmap.installAgentEmptyText"),
+                SimpleTextAttributes.LINK_ATTRIBUTES,
+                e -> UserMilestonesEditorProvider.openUserMilestones(project));
         tree.setRootVisible(false);
         tree.setShowsRootHandles(false);
 
@@ -147,6 +155,11 @@ public class AppMapWindowPanel extends SimpleToolWindowPanel implements DataProv
         new EditSourceOnDoubleClickHandler.TreeMouseListener(tree, null).installOn(tree);
         EditSourceOnEnterKeyHandler.install(tree);
         return tree;
+    }
+
+    @NotNull
+    private JPanel createUserMilestonesPanel() {
+        return new UserMilestonesPanel(project);
     }
 
     public void rebuild(boolean force) {